feat: harden multilingual ASR segment processing
Normalize detected languages before alignment, prevent ASR padding from crossing speaker boundaries, add coverage and document both OpenAI and AsyncOpenAI clients.

+++ app/services/serialization.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 from collections import defaultdict
 from typing import Any
 
+import numpy as np
+
 from app.domain.embeddings import WeightedEmbedding, aggregate_embeddings
 
 
@@ -13,8 +15,6 @@ def build_speaker_profiles(
     for segment in segments:
         embedding = segment.get("speaker_embedding")
         if embedding:
-            import numpy as np
-
             grouped[segment["speaker"]].append(
                 WeightedEmbedding(
                     vector=np.asarray(embedding, dtype=np.float32),
@@ -41,7 +41,7 @@ def build_speaker_profiles(
 
 def to_srt(segments: list[dict[str, Any]]) -> str:
     def timestamp(value: float) -> str:
-        milliseconds = int(round(value * 1000))
+        milliseconds = round(value * 1000)
         hours, remainder = divmod(milliseconds, 3_600_000)
         minutes, remainder = divmod(remainder, 60_000)
         seconds, millis = divmod(remainder, 1000)
@@ -58,7 +58,7 @@ def to_srt(segments: list[dict[str, Any]]) -> str:
 
 def to_vtt(segments: list[dict[str, Any]]) -> str:
     def timestamp(value: float) -> str:
-        milliseconds = int(round(value * 1000))
+        milliseconds = round(value * 1000)
         hours, remainder = divmod(milliseconds, 3_600_000)
         minutes, remainder = divmod(remainder, 60_000)
         seconds, millis = divmod(remainder, 1000)
